Rename project level member/token related files
