+++ webapp/src/dogma/features/metadata/AppMemberDto.ts
@@ -1,0 +1,7 @@
+export type AppMemberDto = Map<string, AppMemberDetailDto>;
+
+export interface AppMemberDetailDto {
+  login: string;
+  role: string;
+  creation: { user: string; timestamp: string };
+}
