+++ webapp/src/dogma/features/metadata/AppMemberList.tsx
@@ -4,21 +4,21 @@ import { DateWithTooltip } from 'dogma/common/components/DateWithTooltip';
 import { UserRole } from 'dogma/common/components/UserRole';
 import { DataTableClientPagination } from 'dogma/common/components/table/DataTableClientPagination';
 import { useDeleteMemberMutation } from 'dogma/features/api/apiSlice';
-import { DeleteMember } from 'dogma/features/repo/DeleteMember';
-import { RepoMemberDetailDto } from 'dogma/features/repo/RepoMemberDto';
+import { DeleteMember } from 'dogma/features/metadata/DeleteMember';
+import { AppMemberDetailDto } from 'dogma/features/metadata/AppMemberDto';
 import { useMemo } from 'react';
 
-export type RepoMemberListProps<Data extends object> = {
+export type AppMemberListProps<Data extends object> = {
   data: Data[];
   projectName: string;
 };
 
-const RepoMemberList = <Data extends object>({ data, projectName }: RepoMemberListProps<Data>) => {
+const AppMemberList = <Data extends object>({ data, projectName }: AppMemberListProps<Data>) => {
   const [deleteMember, { isLoading }] = useDeleteMemberMutation();
-  const columnHelper = createColumnHelper<RepoMemberDetailDto>();
+  const columnHelper = createColumnHelper<AppMemberDetailDto>();
   const columns = useMemo(
     () => [
-      columnHelper.accessor((row: RepoMemberDetailDto) => row.login, {
+      columnHelper.accessor((row: AppMemberDetailDto) => row.login, {
         cell: (info) => (
           <VStack alignItems="left">
             <Text>{info.getValue()}</Text>
@@ -29,15 +29,15 @@ const RepoMemberList = <Data extends object>({ data, projectName }: RepoMemberLi
         ),
         header: 'Login ID',
       }),
-      columnHelper.accessor((row: RepoMemberDetailDto) => row.creation.user, {
+      columnHelper.accessor((row: AppMemberDetailDto) => row.creation.user, {
         cell: (info) => info.getValue(),
         header: 'Added By',
       }),
-      columnHelper.accessor((row: RepoMemberDetailDto) => row.creation.timestamp, {
+      columnHelper.accessor((row: AppMemberDetailDto) => row.creation.timestamp, {
         cell: (info) => <DateWithTooltip date={info.getValue()} />,
         header: 'Added At',
       }),
-      columnHelper.accessor((row: RepoMemberDetailDto) => row.login, {
+      columnHelper.accessor((row: AppMemberDetailDto) => row.login, {
         cell: (info) => (
           <DeleteMember
             projectName={projectName}
@@ -55,4 +55,4 @@ const RepoMemberList = <Data extends object>({ data, projectName }: RepoMemberLi
   return <DataTableClientPagination columns={columns as ColumnDef<Data>[]} data={data} />;
 };
 
-export default RepoMemberList;
+export default AppMemberList;
