+++ webapp/src/dogma/features/metadata/AppTokenList.tsx
@@ -3,22 +3,22 @@ import { ColumnDef, createColumnHelper } from '@tanstack/react-table';
 import { DateWithTooltip } from 'dogma/common/components/DateWithTooltip';
 import { UserRole } from 'dogma/common/components/UserRole';
 import { DataTableClientPagination } from 'dogma/common/components/table/DataTableClientPagination';
-import { RepoTokenDetailDto } from 'dogma/features/repo/RepoTokenDto';
+import { AppTokenDetailDto } from 'dogma/features/metadata/AppTokenDto';
 import { useMemo } from 'react';
 import { useDeleteTokenMemberMutation } from 'dogma/features/api/apiSlice';
-import { DeleteMember } from 'dogma/features/repo/DeleteMember';
+import { DeleteMember } from 'dogma/features/metadata/DeleteMember';
 
-export type RepoTokenListProps<Data extends object> = {
+export type AppTokenListProps<Data extends object> = {
   data: Data[];
   projectName: string;
 };
 
-const AppTokenList = <Data extends object>({ data, projectName }: RepoTokenListProps<Data>) => {
+const AppTokenList = <Data extends object>({ data, projectName }: AppTokenListProps<Data>) => {
   const [deleteMember, { isLoading }] = useDeleteTokenMemberMutation();
-  const columnHelper = createColumnHelper<RepoTokenDetailDto>();
+  const columnHelper = createColumnHelper<AppTokenDetailDto>();
   const columns = useMemo(
     () => [
-      columnHelper.accessor((row: RepoTokenDetailDto) => row.appId, {
+      columnHelper.accessor((row: AppTokenDetailDto) => row.appId, {
         cell: (info) => (
           <VStack alignItems="left">
             <Text>{info.getValue()}</Text>
@@ -29,15 +29,15 @@ const AppTokenList = <Data extends object>({ data, projectName }: RepoTokenListP
         ),
         header: 'App ID',
       }),
-      columnHelper.accessor((row: RepoTokenDetailDto) => row.creation.user, {
+      columnHelper.accessor((row: AppTokenDetailDto) => row.creation.user, {
         cell: (info) => info.getValue(),
         header: 'Added By',
       }),
-      columnHelper.accessor((row: RepoTokenDetailDto) => row.creation.timestamp, {
+      columnHelper.accessor((row: AppTokenDetailDto) => row.creation.timestamp, {
         cell: (info) => <DateWithTooltip date={info.getValue()} />,
         header: 'Added At',
       }),
-      columnHelper.accessor((row: RepoTokenDetailDto) => row.appId, {
+      columnHelper.accessor((row: AppTokenDetailDto) => row.appId, {
         cell: (info) => (
           <DeleteMember
             projectName={projectName}
