+++ webapp/src/dogma/features/metadata/DeleteMemberDto.ts
@@ -1,4 +1,4 @@
-export interface DeleteRepoMemberDto {
+export interface DeleteMemberDto {
   projectName: string;
   id: string;
 }
