+++ webapp/src/dogma/features/metadata/NewAppToken.tsx
@@ -2,8 +2,6 @@ import {
   Button,
   FormControl,
   FormErrorMessage,
-  FormLabel,
-  Input,
   Popover,
   PopoverArrow,
   PopoverBody,
@@ -18,19 +16,34 @@ import {
   Stack,
   useDisclosure,
 } from '@chakra-ui/react';
-import { useForm } from 'react-hook-form';
+import { Controller, useForm } from 'react-hook-form';
 import { IoMdArrowDropdown } from 'react-icons/io';
 import { useState } from 'react';
-import { ConfirmAddMember } from 'dogma/features/repo/ConfirmAddMember';
+import { useGetTokensQuery } from 'dogma/features/api/apiSlice';
+import { TokenDto } from 'dogma/features/token/TokenDto';
+import { OptionBase, Select } from 'chakra-react-select';
+import { ConfirmAddToken } from 'dogma/features/metadata/ConfirmAddToken';
+
+interface TokenOptionType extends OptionBase {
+  value: string;
+  label: string;
+}
 
 type FormData = {
-  id: string;
+  appId: string;
   role: string;
 };
 
-export const NewMember = ({ projectName }: { projectName: string }) => {
+export const NewAppToken = ({ projectName }: { projectName: string }) => {
+  const result = useGetTokensQuery();
+  const tokenOptions: TokenOptionType[] = (result.data || [])
+    .filter((token: TokenDto) => !token.deactivation)
+    .map((token: TokenDto) => ({
+      value: token.appId,
+      label: token.appId,
+    }));
   const {
-    register,
+    control,
     handleSubmit,
     reset,
     formState: { errors },
@@ -41,20 +54,20 @@ export const NewMember = ({ projectName }: { projectName: string }) => {
     onToggle: onConfirmAddToggle,
     onClose: onConfirmAddClose,
   } = useDisclosure();
-  const [id, setId] = useState('');
+  const [appId, setAppId] = useState('');
   const [role, setRole] = useState('member');
   const onSubmit = async (data: FormData) => {
-    setId(data.id);
+    setAppId(data.appId);
     onConfirmAddToggle();
   };
   return (
     <Popover placement="bottom" isOpen={isOpen} onClose={onClose}>
       <PopoverTrigger>
         <Button colorScheme="teal" size="sm" onClick={onToggle} rightIcon={<IoMdArrowDropdown />}>
-          New Member
+          New Token
         </Button>
       </PopoverTrigger>
-      <PopoverContent minWidth="max-content">
+      <PopoverContent minWidth="md">
         <PopoverHeader pt={4} fontWeight="bold" border={0} mb={3}>
           Project {projectName}
         </PopoverHeader>
@@ -62,10 +75,29 @@ export const NewMember = ({ projectName }: { projectName: string }) => {
         <PopoverCloseButton />
         <form onSubmit={handleSubmit(onSubmit)}>
           <PopoverBody minWidth="max-content">
-            <FormControl isInvalid={errors.id ? true : false} isRequired>
-              <FormLabel>Login ID</FormLabel>
-              <Input type="text" placeholder="abc123" {...register('id', { required: true })} />
-              {errors.id && <FormErrorMessage>ID is required</FormErrorMessage>}
+            <FormControl isInvalid={errors.appId ? true : false} isRequired>
+              <Controller
+                control={control}
+                name="appId"
+                rules={{ required: true }}
+                render={({ field: { onChange, value, name, ref } }) => (
+                  <Select
+                    ref={ref}
+                    id="appId"
+                    name={name}
+                    options={tokenOptions}
+                    // The default value of React Select must be null (and not undefined)
+                    value={tokenOptions.find((option) => option.value === value) || null}
+                    onChange={(option) => option && onChange(option.value)}
+                    placeholder="Enter App ID ..."
+                    closeMenuOnSelect={true}
+                    openMenuOnFocus={true}
+                    isSearchable={true}
+                    isClearable={true}
+                  />
+                )}
+              />
+              {errors.appId && <FormErrorMessage>App ID is required</FormErrorMessage>}
             </FormControl>
             <RadioGroup defaultValue="member" mt={3} onChange={setRole} value={role}>
               <Stack spacing={5} direction="row">
@@ -80,9 +112,9 @@ export const NewMember = ({ projectName }: { projectName: string }) => {
           </PopoverBody>
           <PopoverFooter border="0" display="flex" alignItems="center" justifyContent="space-between" pb={4}>
             <Spacer />
-            <ConfirmAddMember
+            <ConfirmAddToken
               projectName={projectName}
-              id={id}
+              id={appId}
               role={role}
               isOpen={isConfirmAddOpen}
               onClose={onConfirmAddClose}
