+++ webapp/src/pages/app/projects/metadata/[projectName]/index.tsx
@@ -2,16 +2,16 @@ import { Box, Flex, Heading, Spacer, Tab, TabList, TabPanel, TabPanels, Tabs } f
 import { NewRepo } from 'dogma/features/repo/NewRepo';
 import { Breadcrumbs } from 'dogma/common/components/Breadcrumbs';
 import { useGetMetadataByProjectNameQuery } from 'dogma/features/api/apiSlice';
-import RepoMemberList from 'dogma/features/repo/RepoMemberList';
+import AppMemberList from 'dogma/features/metadata/AppMemberList';
 import RepoPermissionList from 'dogma/features/repo/RepoPermissionList';
-import AppTokenList from 'dogma/features/repo/AppTokenList';
+import AppTokenList from 'dogma/features/metadata/AppTokenList';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import RepoMetaList from 'dogma/features/repo/RepoMetaList';
+import RepoMetaList from 'dogma/features/metadata/RepoMetaList';
 import { DeleteProject } from 'dogma/features/project/DeleteProject';
-import { NewMember } from 'dogma/features/repo/NewMember';
-import { NewAppToken } from 'dogma/features/repo/NewAppToken';
+import { NewMember } from 'dogma/features/metadata/NewMember';
+import { NewAppToken } from 'dogma/features/metadata/NewAppToken';
 
 const tabs = ['repositories', 'permissions', 'members', 'tokens', 'mirror'];
 
@@ -77,7 +77,7 @@ const ProjectMetadataPage = () => {
               <Spacer />
               <NewMember projectName={projectName} />
             </Flex>
-            <RepoMemberList
+            <AppMemberList
               data={metadata ? Array.from(Object.values(metadata.members)) : []}
               projectName={projectName}
             />
